pyproject.toml
[project]
name = "projecthub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.7.1",
    "email-validator>=1.1.3",
    "fastapi>=0.68.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.1",
    "pydantic>=1.8.2",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.5",
    "sqlalchemy>=1.4.23",
    "uvicorn>=0.15.0",
    "websockets>=10.0",
]
